requirements.txt
streamlit
pandas
numpy
plotly
yfinance
requests
python-dotenv
feedparser
